% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAVAKnock.R
\docType{data}
\name{Example.Ghostknockoff-generation}
\alias{Example.Ghostknockoff-generation}
\alias{Ghostknockoff_generation.example}
\title{Data example of LAVA-Knock Z-scores with Ghostknockoff generation.}
\format{
An object of class \code{list} of length 5.
}
\usage{
data("Ghostknockoff_generation.example")
}
\description{
This example dataset contains the Z-scores for two phenotypes and the genotype matrix of a unit: chr1:58387484-70992905. Due to computational and memory size constaints, we use 100 snps as an example to generation knockoff Z-scores using Ghostknockoff.
}
\examples{
data(Ghostknockoff_generation.example)
chr=Ghostknockoff_generation.example$chr
unit_start=Ghostknockoff_generation.example$unit_start
unit_end=Ghostknockoff_generation.example$unit_end
print(paste0('chr',chr,':',unit_start,'-',unit_end))
#"chr1:58387484-70992905"
Zscore_0_unit=Ghostknockoff_generation.example$Zscore_0_unit
dim(Zscore_0_unit)
G_unit=Ghostknockoff_generation.example$G_unit
dim(G_unit)
}
\keyword{data}
