% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAVAKnock.R
\docType{data}
\name{Example.LAVAKnock}
\alias{Example.LAVAKnock}
\alias{Results_LAVAKnock.example}
\title{Data example of LAVA-Knock knockoff filter for multiple testing.}
\format{
An object of class \code{data.frame} with 1468 rows and 7 columns.
}
\usage{
data("Results_LAVAKnock.example")
}
\description{
This example dataset contains the LAVAKnock bivariateresults on 1,468 windows. Each window has an original p-values and M=5 knockoff p-values
}
\examples{
data("Results_LAVAKnock.example")
dim(Results_LAVAKnock.example) #1468    7
}
\keyword{data}
