% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAVAKnock.R
\docType{data}
\name{Example.LAVAKnock_univariate}
\alias{Example.LAVAKnock_univariate}
\alias{LAVAKnock_univariate.example}
\title{Data example for LAVA-Knock univariate heritability test on one locus.}
\format{
An object of class \code{list} of length 3.
}
\usage{
data("LAVAKnock_univariate.example")
}
\description{
This example dataset contains a locus on chr1:64,344,227-65,894,184, its genotype matrix G_locus with n=20000 samples and p=197 variants, with the original Z-score of 197 variants and 2 phenotypes
}
\examples{
data("LAVAKnock_univariate.example")

Zscore_0=LAVAKnock_univariate.example$Zscore_0
G_locus=LAVAKnock_univariate.example$G_locus
locus=LAVAKnock_univariate.example$locus
print(locus)
dim(G_locus) #20000  197
dim(Zscore_0) #197   2
}
\keyword{data}
