% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAVAKnock.R
\name{Ghostknockoff_generation}
\alias{Ghostknockoff_generation}
\title{Conduct LAVA-Knock Ghostknockoff generation for a unit.}
\usage{
Ghostknockoff_generation(G_unit, Zscore_0_unit, LD.threshold = 0.75, n = 20000)
}
\arguments{
\item{G_unit}{The genotype matrix or the reference genotype for a unit, a n*p matrix}

\item{Zscore_0_unit}{Zscores of a unit, a p*2 matrix for two phenotypes}

\item{LD.threshold}{LD.threshold for single-linkage hierarchical clustering to filter out highly correlated variants. The recommended level is 0.75.}

\item{n}{sample size.}
}
\value{
\item{Zscore_pheno1}{5 knockoff Zscores for phenotype 1.}

\item{Zscore_pheno2}{5 knockoff Zscores for phenotype 2.}
}
\description{
This function generate the knockoff Z-scores for a unit using Ghostknockoff with (shrinkage) empirical LD matrix.
}
\examples{
data(Ghostknockoff_generation.example)
Zscore_0_unit=Ghostknockoff_generation.example$Zscore_0_unit
dim(Zscore_0_unit)
G_unit=Ghostknockoff_generation.example$G_unit
dim(G_unit)

Results_Ghostknockoff_generation=Ghostknockoff_generation(G_unit,
Zscore_0_unit,LD.threshold=0.75,n=20000)
dim(Results_Ghostknockoff_generation$Zscore_pheno1)
dim(Results_Ghostknockoff_generation$Zscore_pheno2)

}
