% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAVAKnock.R
\name{LAVAKnock_bivariate}
\alias{LAVAKnock_bivariate}
\title{Conduct window-level LAVA-Knock bivariate local genetic correlation analysis.}
\usage{
LAVAKnock_bivariate(
  Zscore_pheno1_window = Zscore_pheno1_window,
  Zscore_pheno2_window = Zscore_pheno2_window,
  G_window = G_window,
  chr = chr,
  window_start,
  window_end,
  n = 20000,
  prune.thresh = 99
)
}
\arguments{
\item{Zscore_pheno1_window}{The original and M=5 knockoffs Z-scores of a window under phenotype 1, which is a p*6 matrix for p variants.}

\item{Zscore_pheno2_window}{The original and M=5 knockoffs Z-scores of a window under phenotype 2, which is a p*6 matrix for p variants.}

\item{G_window}{The genotype matrix for a window, which is a n*p matrix.}

\item{chr}{chromosome.}

\item{window_start}{window start.}

\item{window_end}{window end.}

\item{n}{sample size.}

\item{prune.thresh}{Pruning threshold of sungular value decomposition. The recommended level is to explain 99 percent of total variance.}
}
\value{
\item{rg.orginal}{local genetic correlation estimate of original Z-score.}

\item{pval.orginal}{bivariate p-value of local genetic correlation test for original Z-score.}

\item{rg.knockoff}{local genetic correlation estimate of knockoff Z-score.}

\item{pval.knockoff}{bivariate p-value of local genetic correlation test for knockoff Z-score.}
}
\description{
This function conducts the window-level bivariate local genetic correlation test on one window.
}
\examples{
data("LAVAKnock_bivariate.example")
window=LAVAKnock_bivariate.example$window
G_window=LAVAKnock_bivariate.example$G_window
Zscore_pheno1_window=LAVAKnock_bivariate.example$Zscore_pheno1_window
Zscore_pheno2_window=LAVAKnock_bivariate.example$Zscore_pheno2_window
chr=window$chr
window_start=window$start
window_end=window$end
Results_LAVAKnock_bivariate_window=LAVAKnock_bivariate(Zscore_pheno1_window=Zscore_pheno1_window,
Zscore_pheno2_window=Zscore_pheno2_window,
G_window=G_window,chr=chr,window_start,window_end,n=20000,prune.thresh=99)
Results_LAVAKnock_bivariate_window

}
