% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAVAKnock.R
\name{LAVAKnock_univariate}
\alias{LAVAKnock_univariate}
\title{Conduct loci-level univariate heritability test.}
\usage{
LAVAKnock_univariate(
  Zscore = Zscore_0,
  G_locus = G_locus,
  chr,
  locus_start,
  locus_end,
  n = 20000,
  prune.thresh = 99
)
}
\arguments{
\item{Zscore}{The original Z-score of a locus, which is a p*2 matrix for p variants and 2 phenotypes.}

\item{G_locus}{The genotype matrix or the reference genotype for a locus, which is a n*p matrix.}

\item{chr}{chromosome.}

\item{locus_start}{locus start.}

\item{locus_end}{locus end.}

\item{n}{sample size.}

\item{prune.thresh}{Pruning threshold of sungular value decomposition. The recommended level is to explain 99 percent of total variance.}
}
\value{
\item{h2_phe1.orginal}{heritability estimate of phenotype 1.}

\item{h2_phe2.orginal}{heritability estimate of phenotype 2.}

\item{p_uni_phe1.orginal}{p-value of univariate heritability test of phenotype 1.}

\item{p_uni_phe2.orginal}{p-value of univariate heritability test of phenotype 2.}
}
\description{
This function conducts the loci-level univariate heritability test on one locus.
}
\examples{
data("LAVAKnock_univariate.example")
Zscore_0=LAVAKnock_univariate.example$Zscore_0
G_locus=LAVAKnock_univariate.example$G_locus
locus=LAVAKnock_univariate.example$locus
chr=locus$chr
locus_start=locus$start
locus_end=locus$stop
Results_univariate_locus=LAVAKnock_univariate(Zscore=Zscore_0,
G_locus=G_locus,chr,locus_start,locus_end,n=20000,prune.thresh=99)
Results_univariate_locus

}
